Install CNI in all nodes in the cluster
Currently, the prepare_node.sh script received an input of the
number of nodes to deploy CNI to. It also relied on a prefix that
when changed would break the script.

By iterating the list of nodes, and installing CNI on all of them,
the resiliency of the solution is improved. Also, we can start
supporting more than 9 nodes.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>